+++ controllers/user.controller.js
@@ -1,0 +1,5 @@
+var db = require('../db');
+
+module.exports.getLogin = (req,res)=>{
+    res.render('user/login');
+}
