+++ index.js
@@ -7,6 +7,7 @@ app.use(bodyParser.json()) // for parsing application/json
 app.use(bodyParser.urlencoded({ extended: true })) // for parsing application/x-www-form-urlencoded
 
 var indexRoute = require('./routes/index.route');
+var userRoute = require('./routes/user.route');
 
 app.set('view engine','pug');
 app.set('views','./views');
@@ -14,6 +15,7 @@ app.set('views','./views');
 app.use(express.static('public'));
 
 app.use('/',indexRoute);
+app.use('/user',userRoute);
 
 app.listen(port,()=>{
     console.log('Sever listening on port'+port);
