+++ public/style.css
@@ -351,3 +351,4 @@ body{
     padding-top: 10px;
     padding-left: 20px;
 }
+
