+++ routes/index.route.js
@@ -1,7 +1,9 @@
 var express = require('express');
 var router = express.Router();
 var controller = require('../controllers/index.controller')
+var validate = require('../validate/index.validate');
 
+//code for index page
 router.get('/', controller.getIndex)
 
 router.get('/search',controller.getSearch);
@@ -10,6 +12,6 @@ router.get('/create-user', controller.getCreate_user);
 
 router.get('/:id', controller.getView_user);
 
-router.post('/create-user', controller.postCreate_user);
+router.post('/create-user',validate.postCreate, controller.postCreate_user);
 
 module.exports = router;
