+++ views/user/login.pug
@@ -1,0 +1,29 @@
+doctype html
+html
+    head
+        link(rel='stylesheet' type='text/css' href='../style.css')
+        link(rel='stylesheet' href='https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css' integrity='sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm' crossorigin='anonymous')
+    body
+        #nav
+            ul.nav-menu
+                li.menu-item.brand My contact
+                li.menu-item
+                    a(href='/') Home
+                li.menu-item
+                    a(href='/create-user') Create User
+                li.menu-item
+                    a(href='/user/login' role="button") Login 
+        .container(style='background-color: rgba(0,43,255,0.1); height: 600px; width: 600px;')
+            h1(style='text-align: center; padding-top: 80px') Login
+            form(style='padding-top: 50px')
+                .form-group
+                    label(for='exampleInputEmail1') Email address
+                    input#exampleInputEmail1.form-control(type='email' aria-describedby='emailHelp' placeholder='Enter email')
+                    small#emailHelp.form-text.text-muted We&apos;ll never share your email with anyone else.
+                .form-group
+                    label(for='exampleInputPassword1') Password
+                    input#exampleInputPassword1.form-control(type='password' placeholder='Password')
+                .form-check
+                    input#exampleCheck1.form-check-input(type='checkbox')
+                    label.form-check-label(for='exampleCheck1') Remeber me
+                button.btn.btn-primary(type='Button') Login
